+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4784,7 +4784,7 @@ export const AgentToolInputRunTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KAD4TRAENBG3TXYP990STGQZ"),
+  id: z.string().default("01KADNRA9VKKX5C3W2S5DMJ537"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$inboundSchema),
@@ -4803,7 +4803,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KAD4TRAENBG3TXYP990STGQZ"),
+  id: z.string().default("01KADNRA9VKKX5C3W2S5DMJ537"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),