+++ packages/orq-rc/RELEASES.md
@@ -5129,3 +5129,13 @@ Based on:
 - [typescript v4.0.0-rc.46] packages/orq-rc
 ### Releases
 - [NPM v4.0.0-rc.46] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.46 - packages/orq-rc
+
+## 2025-11-19 09:06:50
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.658.2 (2.755.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.0.0-rc.48] packages/orq-rc
+### Releases
+- [NPM v4.0.0-rc.48] https://www.npmjs.com/package/@orq-ai/node/v/4.0.0-rc.48 - packages/orq-rc